refactor(frontend): split auth/downloads/imports/discover/notifications api mixins
Move the auth/account, downloads, playlist-import, discover-weekly, and
notification/search domain methods out of api.ts into typed mixin modules.
Pure structural move.

Refactor slice P34 (api-client-typing-split).
