+++ frontend/lib/api.ts
@@ -1,12 +1,12 @@
-import {
-    ApiClientCore,
-    IMPORT_PREVIEW_TIMEOUT_MS,
-    toSearchParams,
-    type ApiData,
-} from "./api/core";
+import { ApiClientCore, type ApiData } from "./api/core";
+import { WithAuth } from "./api/auth";
 import { WithConnectors } from "./api/connectors";
+import { WithDiscover } from "./api/discover";
+import { WithDownloads } from "./api/downloads";
+import { WithImports } from "./api/imports";
 import { WithLibrary } from "./api/library";
 import { WithMedia } from "./api/media";
+import { WithNotifications } from "./api/notifications";
 import { WithPlaylists } from "./api/playlists";
 import { WithPlays } from "./api/plays";
 import { WithRecommendations } from "./api/recommendations";
@@ -360,494 +360,15 @@ export interface PlaybackClientMetricInput {
     fields?: Record<string, unknown>;
 }
 
-class ApiClient extends WithConnectors(WithSettings(WithPlays(WithRecommendations(WithMedia(
-    WithPlaylists(WithLibrary(ApiClientCore))
+class ApiClient extends WithNotifications(WithDiscover(WithImports(WithDownloads(WithAuth(
+    WithConnectors(WithSettings(WithPlays(WithRecommendations(WithMedia(
+        WithPlaylists(WithLibrary(ApiClientCore))
+    )))))
 ))))) {
-    // Onboarding
-    async getOnboardingStatus(): Promise<{
-        needsOnboarding: boolean;
-        hasAccount: boolean;
-    }> {
-        return this.request<{
-            needsOnboarding: boolean;
-            hasAccount: boolean;
-        }>("/onboarding/status", { silent404: true });
-    }
-
-    // Auth
-    async login(username: string, password: string, token?: string): Promise<{
-        id: string;
-        username: string;
-        displayName?: string | null;
-        role: string;
-        requires2FA?: boolean;
-        onboardingComplete?: boolean;
-    }> {
-        const data = await this.request<{
-            token?: string;
-            refreshToken?: string;
-            user?: {
-                id: string;
-                username: string;
-                displayName?: string | null;
-                role: string;
-                requires2FA?: boolean;
-                onboardingComplete?: boolean;
-            };
-            id?: string;
-            username?: string;
-            displayName?: string | null;
-            role?: string;
-            requires2FA?: boolean;
-            onboardingComplete?: boolean;
-        }>("/auth/login", {
-            method: "POST",
-            body: JSON.stringify({ username, password, token }),
-        });
-
-        // If login returned JWT tokens, store them
-        if (data.token) {
-            this.setToken(data.token, data.refreshToken);
-        }
-
-        // Return user data in consistent format
-        if (data.user) {
-            return data.user;
-        }
-        return {
-            id: data.id || "",
-            username: data.username || "",
-            displayName: data.displayName,
-            role: data.role || "",
-            requires2FA: data.requires2FA,
-            onboardingComplete: data.onboardingComplete,
-        };
-    }
-
-    async register(fields: {
-        inviteCode: string;
-        username: string;
-        displayName: string;
-        password: string;
-        confirmPassword: string;
-        email: string;
-    }) {
-        const data = await this.request<{
-            token: string;
-            refreshToken: string;
-            user: {
-                id: string;
-                username: string;
-                displayName: string | null;
-                role: string;
-            };
-        }>("/auth/register", {
-            method: "POST",
-            body: JSON.stringify(fields),
-        });
-        // Store tokens on success
-        if (data.token) {
-            this.setToken(data.token, data.refreshToken);
-        }
-        return data;
-    }
-
-    async createInviteCode(ttl: string, maxUses?: number) {
-        return this.request<{
-            id: string;
-            code: string;
-            expiresAt: string | null;
-            maxUses: number;
-            createdAt: string;
-        }>("/auth/invite-codes", {
-            method: "POST",
-            body: JSON.stringify({ ttl, maxUses: maxUses ?? 1 }),
-        });
-    }
-
-    async getInviteCodes() {
-        return this.request<
-            Array<{
-                id: string;
-                code: string;
-                status: "active" | "expired" | "exhausted" | "revoked";
-                maxUses: number;
-                useCount: number;
-                expiresAt: string | null;
-                createdAt: string;
-                createdBy: string;
-            }>
-        >("/auth/invite-codes");
-    }
-
-    async revokeInviteCode(id: string) {
-        return this.request<{ message: string }>(`/auth/invite-codes/${id}`, {
-            method: "DELETE",
-        });
-    }
-
-    async logout() {
-        await this.request<void>("/auth/logout", {
-            method: "POST",
-        });
-        // Clear the stored JWT token
-        this.clearToken();
-    }
-
-    async getCurrentUser() {
-        return this.request<{
-            id: string;
-            username: string;
-            displayName?: string | null;
-            role: string;
-            onboardingComplete?: boolean;
-            enrichmentSettings?: { enabled: boolean; lastRun?: string };
-            createdAt: string;
-        }>("/auth/me");
-    }
-
-    async getSubsonicPasswordStatus(): Promise<{ hasPassword: boolean }> {
-        return this.request<{ hasPassword: boolean }>("/auth/subsonic-password");
-    }
-
-    async setSubsonicPassword(password: string): Promise<{ success: boolean }> {
-        return this.request<{ success: boolean }>("/auth/subsonic-password", {
-            method: "POST",
-            body: JSON.stringify({ password }),
-        });
-    }
-
-    async clearSubsonicPassword(): Promise<{ success: boolean }> {
-        return this.request<{ success: boolean }>("/auth/subsonic-password", {
-            method: "DELETE",
-        });
-    }
-
     async getYtMusicMixes(): Promise<{
         mixes: Array<{ playlistId: string; title: string; description: string; thumbnails: Array<{ url: string; width: number }>; count: string | null }>;
     }> {
         return this.get("/browse/ytmusic/mixes");
-    }
-
-    // Downloads (Lidarr)
-    async downloadAlbum(
-        artistName: string,
-        albumTitle: string,
-        rgMbid?: string,
-        downloadType: "library" | "discovery" = "library"
-    ) {
-        return this.request<ApiData>("/downloads", {
-            method: "POST",
-            body: JSON.stringify({
-                type: "album",
-                subject: `${artistName} - ${albumTitle}`,
-                mbid: rgMbid,
-                artistName,
-                albumTitle,
-                downloadType,
-            }),
-        });
-    }
-
-    async downloadArtist(
-        artistName: string,
-        mbid: string,
-        downloadType: "library" | "discovery" = "library"
-    ) {
-        return this.request<ApiData>("/downloads", {
-            method: "POST",
-            body: JSON.stringify({
-                type: "artist",
-                subject: artistName,
-                mbid,
-                downloadType,
-            }),
-        });
-    }
-
-    async getDownloadStatus(id: string) {
-        return this.request<ApiData>(`/downloads/${id}`);
-    }
-
-    async getDownloads(limit?: number, includeDiscovery: boolean = false) {
-        const params = new URLSearchParams();
-        if (limit) params.set("limit", String(limit));
-        params.set("includeDiscovery", String(includeDiscovery));
-        const query = params.toString() ? `?${params.toString()}` : "";
-        return this.request<ApiData[]>(`/downloads${query}`);
-    }
-
-    async getDownloadAvailability() {
-        return this.request<{
-            enabled: boolean;
-            lidarr: boolean;
-            soulseek: boolean;
-            tidal: boolean;
-        }>("/downloads/availability");
-    }
-
-    async previewPlaylistImport(
-        url: string
-    ): Promise<PlaylistImportPreviewResponse> {
-        return this.request("/import/preview", {
-            method: "POST",
-            body: JSON.stringify({ url }),
-            timeoutMs: IMPORT_PREVIEW_TIMEOUT_MS,
-        });
-    }
-
-    async executePlaylistImport(input: {
-        previewData: PlaylistImportPreviewResponse;
-        name?: string;
-    }): Promise<PlaylistImportExecuteResponse> {
-        return this.post("/import/execute", input);
-    }
-
-    async previewM3UImport(
-        content: string,
-        name?: string
-    ): Promise<PlaylistImportPreviewResponse> {
-        return this.request("/import/m3u/preview", {
-            method: "POST",
-            body: JSON.stringify({ content, name }),
-        });
-    }
-
-    async submitImportJob(url: string, name?: string) {
-        return this.request<{
-            deduped: boolean;
-            job: ImportJob;
-        }>("/import/jobs", {
-            method: "POST",
-            body: JSON.stringify({ url, name }),
-        });
-    }
-
-    async listImportJobs() {
-        return this.request<{ jobs: ImportJob[] }>("/import/jobs");
-    }
-
-    async getImportJob(jobId: string) {
-        return this.request<{ job: ImportJob }>(`/import/jobs/${jobId}`);
-    }
-
-    async reconnectImportJob(url: string) {
-        return this.request<{ job: ImportJob }>("/import/jobs/reconnect", {
-            method: "POST",
-            body: JSON.stringify({ url }),
-        });
-    }
-
-    async cancelImportJob(jobId: string) {
-        return this.request<{ job: ImportJob }>(`/import/jobs/${jobId}/cancel`, {
-            method: "POST",
-        });
-    }
-
-    async deleteDownload(id: string) {
-        return this.request<{ success: boolean }>(`/downloads/${id}`, {
-            method: "DELETE",
-        });
-    }
-
-    async getAlbumReleases(
-        albumMbid: string,
-        artistName: string,
-        albumTitle: string
-    ): Promise<{
-        albumMbid: string;
-        lidarrAlbumId: number;
-        releases: AlbumRelease[];
-        total: number;
-    }> {
-        const params = new URLSearchParams({ artistName, albumTitle });
-        return this.request(
-            `/downloads/releases/${albumMbid}?${params.toString()}`
-        );
-    }
-
-    async grabRelease(options: {
-        guid: string;
-        indexerId: number;
-        albumMbid: string;
-        lidarrAlbumId: number;
-        artistName: string;
-        albumTitle: string;
-        title: string;
-    }): Promise<{
-        success: boolean;
-        jobId: string;
-        message: string;
-        duplicate?: boolean;
-    }> {
-        return this.request("/downloads/grab", {
-            method: "POST",
-            body: JSON.stringify(options),
-        });
-    }
-
-    // Discover Weekly
-    async generateDiscoverWeekly() {
-        return this.request<{ message: string; jobId: string }>(
-            "/discover/generate",
-            {
-                method: "POST",
-            }
-        );
-    }
-
-    async getDiscoverGenerationStatus(jobId: string) {
-        return this.request<{
-            status: string;
-            progress: number;
-            result?: {
-                success: boolean;
-                playlistName: string;
-                songCount: number;
-                error?: string;
-            };
-        }>(`/discover/generate/status/${jobId}`);
-    }
-
-    async getCurrentDiscoverWeekly() {
-        return this.request<{
-            weekStart: string;
-            weekEnd: string;
-            tracks: ApiData[];
-            unavailable: ApiData[];
-            totalCount: number;
-            unavailableCount: number;
-        }>("/discover/current");
-    }
-
-    async getDiscoverBatchStatus() {
-        return this.request<{
-            active: boolean;
-            status: "downloading" | "scanning" | "generating" | null;
-            batchId?: string;
-            progress?: number;
-            completed?: number;
-            failed?: number;
-            total?: number;
-        }>("/discover/batch-status");
-    }
-
-    async likeDiscoverAlbum(albumId: string) {
-        return this.request<{ success: boolean }>("/discover/like", {
-            method: "POST",
-            body: JSON.stringify({ albumId }),
-        });
-    }
-
-    async unlikeDiscoverAlbum(albumId: string) {
-        return this.request<{ success: boolean }>("/discover/unlike", {
-            method: "DELETE",
-            body: JSON.stringify({ albumId }),
-        });
-    }
-
-    async getDiscoverConfig() {
-        return this.request<{
-            id: string;
-            userId: string;
-            playlistSize: number;
-            enabled: boolean;
-            lastGeneratedAt: string | null;
-        }>("/discover/config");
-    }
-
-    async updateDiscoverConfig(config: {
-        playlistSize?: number;
-        enabled?: boolean;
-    }) {
-        return this.request<{
-            id: string;
-            userId: string;
-            playlistSize: number;
-            enabled: boolean;
-            lastGeneratedAt: string | null;
-        }>("/discover/config", {
-            method: "PATCH",
-            body: JSON.stringify(config),
-        });
-    }
-
-    async clearDiscoverPlaylist() {
-        return this.request<{
-            success: boolean;
-            message: string;
-            likedMoved: number;
-            activeDeleted: number;
-        }>("/discover/clear", {
-            method: "DELETE",
-        });
-    }
-
-    // Discovery Exclusions
-    async getDiscoverExclusions() {
-        return this.request<{
-            exclusions: Array<{
-                id: string;
-                albumMbid: string;
-                artistName: string;
-                albumTitle: string;
-                lastSuggestedAt: string;
-                expiresAt: string;
-            }>;
-            count: number;
-        }>("/discover/exclusions");
-    }
-
-    async clearDiscoverExclusions() {
-        return this.request<{
-            success: boolean;
-            message: string;
-            clearedCount: number;
-        }>("/discover/exclusions", {
-            method: "DELETE",
-        });
-    }
-
-    async removeDiscoverExclusion(id: string) {
-        return this.request<{
-            success: boolean;
-            message: string;
-        }>(`/discover/exclusions/${id}`, {
-            method: "DELETE",
-        });
-    }
-
-    // Artists (Discovery)
-    async getArtistDiscovery(
-        nameOrMbid: string,
-        options?: {
-            includeDiscography?: boolean;
-            includeTopTracks?: boolean;
-            includeSimilarArtists?: boolean;
-        }
-    ) {
-        const queryString = options
-            ? toSearchParams(options).toString()
-            : "";
-        const suffix = queryString ? `?${queryString}` : "";
-        return this.request<ApiData>(
-            `/artists/discover/${encodeURIComponent(nameOrMbid)}${suffix}`
-        );
-    }
-
-    async getAlbumDiscovery(
-        rgMbid: string,
-        options?: {
-            includeTracks?: boolean;
-        }
-    ) {
-        const queryString = options
-            ? toSearchParams(options).toString()
-            : "";
-        const suffix = queryString ? `?${queryString}` : "";
-        return this.request<ApiData>(
-            `/artists/album/${encodeURIComponent(rgMbid)}${suffix}`
-        );
     }
 
     // Audiobooks
@@ -1026,52 +547,6 @@ class ApiClient extends WithConnectors(WithSettings(WithPlays(WithRecommendation
             {
                 method: "DELETE",
             }
-        );
-    }
-
-    // Search
-    async search(
-        query: string,
-        type:
-            | "all"
-            | "artists"
-            | "albums"
-            | "tracks"
-            | "audiobooks"
-            | "podcasts" = "all",
-        limit: number = 20,
-        signal?: AbortSignal
-    ) {
-        return this.request<ApiData>(
-            `/search?q=${encodeURIComponent(query)}&type=${type}&limit=${limit}`,
-            { signal }
-        );
-    }
-
-    async discoverSearch(
-        query: string,
-        type: "music" | "podcasts" | "all" = "music",
-        limit: number = 20,
-        signal?: AbortSignal
-    ) {
-        return this.request<{
-            results: ApiData[];
-            aliasInfo: { original: string; canonical: string; mbid?: string } | null;
-        }>(
-            `/search/discover?q=${encodeURIComponent(query)}&type=${type}&limit=${limit}`,
-            { signal }
-        );
-    }
-
-    async discoverSimilarArtists(
-        artist: string,
-        mbid: string = "",
-        limit: number = 6,
-        signal?: AbortSignal
-    ) {
-        return this.request<{ similarArtists: ApiData[] }>(
-            `/search/discover/similar?artist=${encodeURIComponent(artist)}&mbid=${encodeURIComponent(mbid)}&limit=${limit}`,
-            { signal }
         );
     }
 
@@ -1369,102 +844,6 @@ class ApiClient extends WithConnectors(WithSettings(WithPlays(WithRecommendation
         return this.request(url);
     }
 
-    // API Keys Management
-    async createApiKey(deviceName: string): Promise<{
-        apiKey: string;
-        name: string;
-        createdAt: string;
-        message: string;
-    }> {
-        return this.post("/api-keys", { deviceName });
-    }
-
-    async listApiKeys(): Promise<{
-        apiKeys: Array<{
-            id: string;
-            name: string;
-            createdAt: string;
-            lastUsed: string | null;
-        }>;
-    }> {
-        return this.get("/api-keys");
-    }
-
-    async revokeApiKey(id: string): Promise<{ message: string }> {
-        return this.delete(`/api-keys/${id}`);
-    }
-
-    async getNotifications(): Promise<
-        Array<{
-            id: string;
-            type: string;
-            title: string;
-            message: string | null;
-            metadata: ApiData | null;
-            read: boolean;
-            cleared: boolean;
-            createdAt: string;
-        }>
-    > {
-        return this.get("/notifications");
-    }
-
-    async getUnreadNotificationCount(): Promise<{ count: number }> {
-        return this.get("/notifications/unread-count");
-    }
-
-    async markNotificationAsRead(id: string): Promise<{ success: boolean }> {
-        return this.post(`/notifications/${id}/read`);
-    }
-
-    async markAllNotificationsAsRead(): Promise<{ success: boolean }> {
-        return this.post("/notifications/read-all");
-    }
-
-    async clearNotification(id: string): Promise<{ success: boolean }> {
-        return this.post(`/notifications/${id}/clear`);
-    }
-
-    async clearAllNotifications(): Promise<{ success: boolean }> {
-        return this.post("/notifications/clear-all");
-    }
-
-    // Download Activity
-    async getActiveDownloads(): Promise<
-        Array<{
-            id: string;
-            subject: string;
-            type: string;
-            status: string;
-            createdAt: string;
-            error?: string;
-        }>
-    > {
-        return this.get("/notifications/downloads/active");
-    }
-
-    async getDownloadHistory(): Promise<
-        Array<{
-            id: string;
-            subject: string;
-            type: string;
-            status: string;
-            error?: string;
-            createdAt: string;
-            completedAt?: string;
-        }>
-    > {
-        return this.get("/notifications/downloads/history");
-    }
-
-    async clearDownloadFromHistory(id: string): Promise<{ success: boolean }> {
-        return this.post(`/notifications/downloads/${id}/clear`);
-    }
-
-    async clearAllDownloadHistory(): Promise<{ success: boolean }> {
-        return this.post("/notifications/downloads/clear-all");
-    }
-
     // Vibe (CLAP Similarity) API
     async getVibeSimilarTracks(trackId: string, limit = 20) {
         return this.request<{
@@ -1660,12 +1039,6 @@ class ApiClient extends WithConnectors(WithSettings(WithPlays(WithRecommendation
                 error?: string;
             }>;
         }>("/podcasts/refresh-all", { method: "POST" });
-    }
-
-    async retryFailedDownload(
-        id: string
-    ): Promise<{ success: boolean; newJobId?: string }> {
-        return this.post(`/notifications/downloads/${id}/retry`);
     }
 
     // ── YouTube Music ──────────────────────────────────────────────
